+++ tests/SupportEnums/MyBackedEnum.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Zumba\JsonSerializer\Test\SupportEnums;
+
+enum MyBackedEnum: string
+{
+    case Hearts = 'H';
+    case Diamonds = 'D';
+    case Clubs = 'C';
+    case Spades = 'S';
+}