+++ tests/SupportEnums/MyUnitEnum.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Zumba\JsonSerializer\Test\SupportEnums;
+
+enum MyUnitEnum
+{
+    case Hearts;
+    case Diamonds;
+    case Clubs;
+    case Spades;
+}