+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support of UnitEnum and BackedEnum deserialization (PHP > 8.1)
 ### Fixed
 - Fixed deprecated with DateTimeImmutable deserialization with PHP 8.2